Kursova/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic-settings==2.2.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1